requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-Bcrypt>=1.0
PyMySQL>=1.1
PyJWT>=2.8
python-dotenv>=1.0
gunicorn>=21.2
cryptography>=41.0
